feat: make two-factor authentication available to all users
Two-factor authentication was originally introduced as a Plus-only feature
in #2565. Security shouldn't be a paid upsell — open it up to community
users too.

- Drop the #[RequiresPlus] attribute from the four 2FA management
  controllers (setup, confirm, disable, regenerate-recovery-codes); the
  challenge endpoint and the credentials login flow were never gated.
- Collapse the Plus-edition feature tests into tests/Feature/Auth/, where
  they cover the same paths against TestCase instead of PlusTestCase, and
  drop the now-moot "managementRoutesAreForbiddenOnCommunity" guard.
