+++ app/Http/Controllers/API/Auth/TwoFactor/SetupController.php
@@ -2,14 +2,12 @@
 
 namespace App\Http\Controllers\API\Auth\TwoFactor;
 
-use App\Attributes\RequiresPlus;
 use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Services\Auth\TwoFactorAuthenticator;
 use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\Response;
 
-#[RequiresPlus]
 class SetupController extends Controller
 {
     public function __construct(
